backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Validation and settings
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# HTTP client for external APIs
httpx==0.26.0

# Caching
redis==5.0.1

# Monitoring
sentry-sdk[fastapi]==1.40.0

# Authentication (bcrypt>=4 breaks passlib 1.7.4 version detection)
passlib[bcrypt]==1.7.4
bcrypt<4.0.0
python-jose[cryptography]==3.3.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
factory-boy==3.3.0

# Code quality (development)
black==24.4.2
isort==5.13.2
ruff==0.4.4
